refactor: move graphql folder into to API config folder
