+++ package.json
@@ -3,13 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "setup:prisma-schema": "yarn run services:up && prisma db push && prisma db seed",
+    "dev": "npm run services:up && next dev",
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
     "prepare": "husky",
-    "compose:up": "docker compose -f ./infra/compose.yaml up -d",
-    "compose:down": "docker compose -f ./infra/compose.yaml down"
+    "services:up": "docker compose -f ./infra/compose.yaml up -d",
+    "services:down": "docker compose -f ./infra/compose.yaml down"
   },
   "dependencies": {
     "@apollo/client": "^3.9.10",
